added some rss feeds

+++ aspects/delete_aspect.php
@@ -2,8 +2,10 @@
 
    include_once("../top.php");
 
-   $delete_id = $_POST['delete_id'];
-   $asp = new Aspect($delete_id);
-   $asp->get('title');
-   $asp->delete();
-   chatline("deletes aspect $title.");
+   if($access == 0){
+       $delete_id = $_POST['delete_id'];
+       $asp = new Aspect($delete_id);
+       $asp->get('title');
+       $asp->delete();
+       chatline("deletes aspect $title.");
+   }
